Convert application to GTK
Now it has a GTK GUI, but is kind of dysfunctional. Mouse movement still
doesn't work (can't get the right events to work). Altered the shaders
to be independent of special handling (e.g., even the Newton polynomial
is now handled as a "kind of" uniform variable). Some minor improvements
here and there (keeping the aspect ratio in the Viewport, for example).
Work in progress.